refactor: move connect logic to component
